+++ src/ccbr_tools/gb2gtf.py
@@ -18,20 +18,24 @@ import Bio
 
 def main():
     if check_args(sys.argv):
-        gb2gtf()
+        gb2gtf(sys.argv)
+
+
+usage_msg = """Convert GenBank files to GTF format.
+
+Usage: gb2gtf sequence.gb > sequence.gtf
+"""
 
 
 def check_args(args):
     valid_usage = True
     if len(args) < 2 or "-h" in args or "--help" in args:
-        print("Usage: gb2gtf sequence.gb > sequence.gtf")
+        print(usage_msg)
         valid_usage = False
     return valid_usage
 
 
-def gb2gtf():
-    args = sys.argv
-
+def gb2gtf(args):
     # get all sequence records for the specified genbank file
     recs = [rec for rec in SeqIO.parse(args[1], "genbank")]
 
@@ -54,7 +58,7 @@ def gb2gtf():
             l = feat.location
             start = l.start
             end = l.end
-            if feat.strand == 1:
+            if feat.location.strand == 1:
                 strand = "+"
             else:
                 strand = "-"