+++ tests/test_templates.py
@@ -3,7 +3,7 @@ import pathlib
 import pytest
 import tempfile
 
-from ccbr_tools.templates import read_template, use_template
+from ccbr_tools.templates import read_template, use_template, use_quarto_ext
 
 
 def test_read_template():
@@ -38,24 +38,6 @@ def test_use_template():
     assert all(assertions)
 
 
-def test_use_template_defaults():
-    with tempfile.TemporaryDirectory() as tmp_dir:
-        current_wd = os.getcwd()
-        tmp_wd = pathlib.Path(current_wd) / tmp_dir
-        os.chdir(tmp_wd)
-        use_template(
-            "submit_slurm.sh",
-            PIPELINE="CCBR_nxf",
-            MODULES="ccbrpipeliner nextflow",
-            ENV_VARS="export HELLO=WORLD",
-            RUN_COMMAND="nextflow run main.nf -stub",
-        )
-        os.chdir(current_wd)
-        template_file = pathlib.Path(tmp_wd) / "submit_slurm.sh"
-        assertions = [template_file.is_file()]
-    assert all(assertions)
-
-
 def test_use_template_blanks():
     with tempfile.TemporaryDirectory() as tmp_dir:
         out_filepath = pathlib.Path(tmp_dir) / "test.sh"
@@ -65,3 +47,23 @@ def test_use_template_blanks():
                 output_filepath=out_filepath,
             )
             assert str(exc_info.value) == "KeyError: 'MODULES'"
+
+
+def test_use_quarto_ext():
+    with tempfile.TemporaryDirectory() as tmp_dir:
+        current_wd = os.getcwd()
+        os.chdir(tmp_dir)
+        use_quarto_ext("fnl")
+        assertions = [
+            (pathlib.Path("_extensions") / "fnl").is_dir(),
+            len(os.listdir(pathlib.Path("_extensions") / "fnl")) > 1,
+            (pathlib.Path("_extensions") / "fnl" / "_extension.yml").is_file(),
+        ]
+        os.chdir(current_wd)
+    assert all(assertions)
+
+
+def test_use_quarto_ext_error():
+    with pytest.raises(FileNotFoundError) as exc_info:
+        use_quarto_ext("not_a_real_extension")
+    assert str(exc_info.value).startswith("not_a_real_extension does not exist")